data table moved into componentDidMount
